backend/ingestion/requirements.txt
openai
numpy
ujson
tqdm
python-dotenv